+++ protocol/graphql/object.go
@@ -4,7 +4,6 @@ import (
 	"errors"
 
 	"github.com/graphql-go/graphql"
-	"github.com/jexia/maestro/refs"
 	"github.com/jexia/maestro/specs"
 	"github.com/jexia/maestro/specs/types"
 )
@@ -46,58 +45,4 @@ func NewObject(name string, prop *specs.Property) (*graphql.Object, error) {
 	}
 
 	return graphql.NewObject(config), nil
-}
-
-// NewArgs construct new field config arguments for the graphql schema
-func NewArgs(specs *specs.Property) (graphql.FieldConfigArgument, error) {
-	args := graphql.FieldConfigArgument{}
-	if specs.Type != types.TypeMessage {
-		return args, nil
-	}
-
-	if len(specs.Nested) == 0 {
-		return nil, nil
-	}
-
-	for _, nested := range specs.Nested {
-		if nested.Type == types.TypeMessage {
-			continue
-		}
-
-		args[nested.Name] = &graphql.ArgumentConfig{
-			Type:        gtypes[nested.Type],
-			Description: nested.Desciptor.GetComment(),
-		}
-	}
-
-	return args, nil
-}
-
-// ResponseValue constructs the response value send back to the client
-func ResponseValue(specs *specs.Property, refs *refs.Store) (interface{}, error) {
-	if specs.Type != types.TypeMessage {
-		return nil, ErrInvalidObject
-	}
-
-	result := make(map[string]interface{}, len(specs.Nested))
-	for _, nested := range specs.Nested {
-		if nested.Type == types.TypeMessage {
-			value, err := ResponseValue(nested, refs)
-			if err != nil {
-				return nil, err
-			}
-
-			result[nested.Name] = value
-			continue
-		}
-
-		ref := refs.Load(nested.Reference.Resource, nested.Reference.Path)
-		if ref == nil {
-			continue
-		}
-
-		result[nested.Name] = ref.Value
-	}
-
-	return result, nil
 }
